Add Deftimeout variable to workaround MIDI input bug from rtmidi
The perfect solution would be to fix mdep_waitfor() so that it properly returned as soon as MIDI input was available.  The default value of Deftimeout is 2 (milliseconds), which means that we'll essentially be polling every 2 milliseconds.  On modern machines, that's not much overhead, but at some point it might be worth fixing this.