Extend LE_3BYTE construct to hold as many bytes as size of NT_NOTE/ON/OFF
Given that the LE_3BYTES note type struct is in a union with the
NT_NOTE/ON/OFF struct, it makes sense to make the LE_3BYTES structure
hold as many bytes to make it the same size as NT_NOTE_ON_OFF struct.
Rename NT_LE3BYTES to be NT_LE_NBYTES since the bytes array is in
a NT_LE_NBYTES is larger than 3.
Update three/twobytes() to create a NT_LE_NBYTES note type instead
of NT_BYTES(which requires a second allocation to hold the separate
Mididata).
Declare array indices affected by this change to be unsigned int.