Add -Debugmemscribble to scribble over allocated memory before use
Modify dbg/allocate to scribble 0xa5 over allocated memory if
newly added Debubmemscribble is set.
Modify bi_sbbyes to set the port of newly allocated not to be that
of midibytes are cloned from.
Add INIT_DEBUGARG_PAIR macro to keylib.c and use to initialized
Debugmemscribble, Debugstrsave and Debugmalloc variables.
Modify keytest.sh to add "-Debugmemscribble" to each lowkey
invocation(and announce what the common command line args are).