Support ALSA port picking
Create new walsaportpick tool that allows specifying which ALSA port to
connect keykit port to.
Extend midi() command (and mdep MIDI) to support four arg "open" where
last arg is the port to connect the kekit port to - and extend Linux
midi if invoked with it to make the ALSA connection. Add "isopento"
midi() arg that returns which port the keykit port is connected to.