Add more branch optimizations
Handle I_FCONDEVAL in multicodeiseg().
Handle case of branching instruction that targets unconditional goto;
update target of branch to that of the goto.
Handle case of condition branch over uncondition goto; invert
condition and update target to that of the goto.
Handle case of not followed by condition branch by inverting
condition, but only if the condition branch is not a target of another
branch.