Fix optimizaiton I that removes branch instructions
In optimization "I" that reverses conditional branch across a goto
isntruction instruction, it failed to not only update the instnode
branch list that the target of the condition branch changed, but also
that the goto instruction was removed. Fix by adding
rminstnodebranch() which removes tracking of a removed branch
instruction, and also call instnodebranchupdate() to update the
target of the conditional branch instruction.